Commit Intellij Workspace configuration
I'm honestly unsure if this is a good idea. But we don't have it in our `.gitignore`, yet. And given that the repo might get quite complex in terms of different languages being included, etc., having a working configuration preconfigured might be useful.

Can be reverted if we don't want to do this.

Refs:
TELESTION-451 Add Gherkin-based executable specifications for Backend Services