Allow developers to enable auto-login for their frontend
With this change, the `setAutoLoginCredentials()` function gets exposed as a public API. While it should be used with extreme care (since using it means hard-coding the NATS credentials into the built frontends), this enables developers to enable auto-login in their Telestion based frontends by calling:

```tsx
setAutoLoginCredentials({
	natsUrl: 'ws://localhost:9222',
	username: 'nats',
	password: 'nats'
});
```

While there are, obviously, security risks associated with this, some use-cases require such functionality which is why it gets exposed as an official API to at least make it as secure as possible by tightly integrating it into the core framework. To make developers aware of the security implications, they are described in the documentation and the function is marked as `@deprecated` for extra attention.