Document the Go module release system (prefixed tags)
Go modules defined in a subdirectory need release tags prefixed with the subdirectory name in order to be recognized [1].
Based on us storing our Go module in the `backend-go` subdirectory of the repo, we therefore need tags prefixed with `backend-go/[release-version]`.
For `v1.0.0-alpha.4` (due to caching problems), this was verified to work with the tag `backend-go/v1.0.0-goalpha.4`, which was found by pkg.go.dev [2].

This commit documents this knowledge (for now inside the Go package `README.md`), until a more formal release workflow documentation / automation gets established.

* [1] https://web.archive.org/web/20240127035037/https://go.dev/ref/mod#vcs-version
* [2] https://pkg.go.dev/github.com/wuespace/telestion/backend-go@v1.0.0-goalpha.4