fix(frontend-react): Fix the export in the `package.json` after the Vite upgrade
With the Vite upgrade to v6, the CSS file got renamed:
<https://vite.dev/guide/migration#customize-css-output-file-name-in-library-mode>.

This means we have to adjust the name in the `exports` of the
`package.json`, as well. This is actually quite nice because the name
now matches the one actually imported. But still: requires a change.
That's what this commit does. Nothing more, nothing less.