Activate footnotes on the documentation site
This is primarily necessary for the migration of the ADRs.
However, it's also nice to have the ability to have footnotes for references and the like.
Therefore, it is useful to add footnotes, which can be achieved using the `footnotes` markdown extension.

M. Donath, “Footnotes - material for MKDocs.” https://squidfunk.github.io/mkdocs-material/reference/footnotes/